requirements.txt
Django
requests~=2.32.4
openai~=1.107.1
djangorestframework~=3.16.1
psycopg2-binary==2.9.10
django-postgres-backup==1.0.0
python-decouple==3.8